Refactor generate_name_description_error_message function to utils module
